# src/models/lessons.ts
import { Lesson } from "@prisma/client";
import { Step } from "./steps"

/**
 * @swagger
 * components:
 *  schemas:
 *    LessonRawDTO:
 *      type: object
 *      properties:
 *        title:
 *          type: string
 *          description: The title of the lesson.
 *          example: Prints
 *        start:
 *          type: string
 *          description: The day-time when the lesson starts. To set the actual date, leave it blank.
 *          example: 2022-10-10T23:37:56.312Z
 *        end:
 *          type: string
 *          description: The day-time when the lesson finishes. If there's no end date, leave it blank.
 *          example: 2022-11-10T23:37:56.312Z
 *        languageName:
 *          type: string
 *          description: The programming language that will be used in the lesson.
 *          example: python
 *      required:
 *        - title
 *        - languageName
 */
export interface LessonRawDTO extends Omit<LessonDTO, 'courseId'>{}

/**
 * @swagger
 * components:
 *  schemas:
 *    LessonDTO:
 *      allOf:
 *          - $ref: '#/components/schemas/LessonRawDTO'
 *          - type: object
 *          - properties:
 *              courseId:
 *                  type: string
 *                  description: The id of the existing course that uses this lesson.
 *                  example: 63158ff83cd164cc4fda4282
 *      required:
 *          - courseId
 */
export interface LessonDTO extends Omit<Lesson, 'id'>{};

/**
 * @swagger
 * components:
 *  schemas:
 *    Lesson:
 *      allOf:
 *        - properties:
 *           id:
 *             type: string
 *             description: The autogenerated id of the lesson.
 *             example: 63158ff83cd164cc4fda4282
 *        - $ref: '#/components/schemas/LessonDTO'
 *        - type: object
 *      required:
 *          - id
 */
export { Lesson };

/**
 * @swagger
 * components:
 *  schemas:
 *    LessonNested:
 *      allOf:
 *        - $ref: '#/components/schemas/Lesson'
 *        - type: object
 *        - properties:
 *           steps:
 *             type: array
 *             items:
 *               $ref: '#/components/schemas/Step'
 *      required:
 *          - steps
 */
export interface LessonNested extends Lesson{
    steps: Step[]
}